GCW-3417 Rename Api client filename
